Restructure WordDetectorNN into integrated word_detector training package. [CC]
