--- xournalpp_htr/training/WordDetectorNN/.python-version
@@ -1,1 +1,0 @@
-3.12
