--- xournalpp_htr/training/WordDetectorNN/pyproject.toml
@@ -1,20 +1,0 @@
-[project]
-name = "worddetectornn"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "dotenv>=0.9.9",
-    "gitpython>=3.1.45",
-    "gradio>=5.47.2",
-    "jupyter>=1.1.1",
-    "opencv-python>=4.12.0.88",
-    "scikit-learn>=1.7.2",
-    "supabase>=2.20.0",
-    "tensorboard>=2.20.0",
-    "torch>=2.7.1",
-    "torchaudio>=2.7.1",
-    "torchvision>=0.22.1",
-    "tqdm>=4.67.1",
-]
