--- xournalpp_htr/training/WordDetectorNN/requirements.txt
@@ -1,10 +1,0 @@
-# For HF Gradio Space
-torch
-torchvision
-numpy
-scikit-learn
-tqdm
-opencv-python
-gitpython
-supabase
-dotenv
