+++ xournalpp_htr/training/word_detector/.env.example
