+++ xournalpp_htr/training/word_detector/gradio-demo-architecture.excalidraw
