+++ xournalpp_htr/training/word_detector/run_training.sh
